+++ src/features/map/UserStationMap.tsx
@@ -43,7 +43,7 @@ export default function UserStationMap(): JSX.Element {
     <MapContainer
       center={[60.171033, 24.941497]}
       zoom={13}
-      style={{ height: '700px', width: '700px' }}
+      style={{ height: '80vh', width: '70vw' }}
     >
       <TileLayer
         attribution='&copy; <a href="https://www.digitransit.fi/en/developers/apis/6-terms-of-use/">Digitransit</a> contributors'